Lock element when inserted
This locks a few elements so that if the documentElement is replaced,
	 these elements will be moved over to the new one.
